Implemented initial version of Hadoop MapReduce AT.
Added new semantic element for that (Isolated Template Completion
Parameters) that are needed if allocation context and/or system are used
from a template (as it is forbidden to have more than 1
system/allocation on the PCM partition).

The model runs through; the results are not satisfactory yet.
Potentially, there is still a little bug either in the AT implementation
or in SimuLizar.