Fixed Daria's code :P
In the AT world, it can happen that there is no resource environment in
the PCM partition. Only after the AT completion ran through, the PCM
partition has to fulfill this constraint.

ALso refactored the code to reduce redundancies and made the list
unmodifiable (why should it be modifiable?!).

Tested with the MapReduce example before commit.